requirements.txt
certifi
GPUtil==1.4.0
psutil==5.9.4
psycopg2-binary==2.9.5
PyYAML==6.0
python-dotenv
